backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
anthropic==0.7.8
pydantic==2.5.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy>=1.26.0
pillow>=10.0.0
ultralytics==8.1.0
torch>=2.0.0
torchvision>=0.15.0
twilio==8.10.0
